+++ routes/adverts.routes.test.js
@@ -4,7 +4,7 @@ const fs = require("fs");
 const path = require("path")
 const { v4: uuidv4 } = require("uuid");
 
-const filePath = path.resolve(__dirname, "../vTwo.adverts.json");
+const filePath = path.resolve(__dirname, "../adverts.json");
 const testFilePath = path.resolve(__dirname, "../adverts.routes.test.data.json");
 
 let simulata = {};
